+++ packages/aws-cloudfront/lib/index.js
@@ -52,11 +52,8 @@ const createCloudFrontDistribution = async (cf, s3, inputs) => {
     } = await createOriginAccessIdentity(cf));
   }
 
-  const originRegion = inputs.originRegion;
-
   const { Origins, CacheBehaviors } = parseInputOrigins(inputs.origins, {
-    originAccessIdentityId,
-    originRegion
+    originAccessIdentityId
   });
 
   if (s3CanonicalUserId) {
@@ -121,11 +118,8 @@ const updateCloudFrontDistribution = async (cf, s3, distributionId, inputs) => {
     } = await createOriginAccessIdentity(cf));
   }
 
-  const originRegion = inputs.originRegion;
-
   const { Origins, CacheBehaviors } = parseInputOrigins(inputs.origins, {
-    originAccessIdentityId,
-    originRegion
+    originAccessIdentityId
   });
 
   if (s3CanonicalUserId) {