+++ packages/lambda-at-edge/src/default-handler.ts
@@ -105,8 +105,9 @@ export const handler = async (
   if (isHTMLPage || isPublicFile) {
     s3Origin.path = isHTMLPage ? "/static-pages" : "/public";
 
+    addS3HostHeader(request, normalisedS3DomainName);
+
     if (isHTMLPage) {
-      addS3HostHeader(request, normalisedS3DomainName);
       request.uri = `${uri}.html`;
     }