+++ packages/aws-cloudfront/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.1](https://github.com/serverless-nextjs/serverless-next.js/compare/@sls-next/aws-cloudfront@1.0.1-alpha.0...@sls-next/aws-cloudfront@1.0.1) (2020-07-11)
+
+**Note:** Version bump only for package @sls-next/aws-cloudfront
+
 ## 1.0.1-alpha.0 (2020-07-05)
 
 **Note:** Version bump only for package @sls-next/aws-cloudfront